feat: removed duplicate view profile page code
