+++ lib/main.dart
@@ -32,7 +32,7 @@ class MyApp extends StatelessWidget {
           theme: regularTheme,
           darkTheme: darkTheme,
           themeMode: themeMode,
-          initialRoute: Routes.loginView,
+          initialRoute: Routes.splashview,
         ),
       ),
     );
